+++ src/main/java/io/github/kreiseljustus/asmputils/Constants.java
@@ -1,6 +1,6 @@
 package io.github.kreiseljustus.asmputils;
 
 public class Constants {
-    static final String VERSION = "1.3.0";
-    static final String VERSION_URL = "https://kreiseljustus.com/asmp_version.txt";
+    public static final String VERSION = "1.3.0";
+    public static final String VERSION_URL = "https://kreiseljustus.com/asmp_version.txt";
 }
