+++ src/main/java/io/github/kreiseljustus/asmputils/core/IModule.java
@@ -3,6 +3,8 @@ package io.github.kreiseljustus.asmputils.core;
 import net.minecraft.util.math.ChunkPos;
 
 public interface IModule {
+    abstract void onInitClient();
+
     abstract void onTick();
 
     abstract void onChunkEnter(ChunkPos chunkPos);
