+++ src/main/java/io/github/kreiseljustus/asmputils/core/VersionManagement.java
@@ -1,5 +1,6 @@
-package io.github.kreiseljustus.asmputils;
+package io.github.kreiseljustus.asmputils.core;
 
+import io.github.kreiseljustus.asmputils.Constants;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.text.Text;
