+++ src/main/java/io/github/kreiseljustus/asmputils/core/data/ShopDataHolder.java
@@ -1,15 +1,9 @@
-package io.github.kreiseljustus.asmputils;
+package io.github.kreiseljustus.asmputils.core.data;
 
 import org.jetbrains.annotations.NotNull;
 
 import java.util.Arrays;
 import java.util.Objects;
-
-class ShopException extends Exception {
-    ShopException() {
-        super("Failed to create shop data");
-    }
-}
 
 public class ShopDataHolder {
     public String Owner;
