+++ src/main/java/io/github/kreiseljustus/asmputils/core/data/ShopException.java
@@ -1,0 +1,7 @@
+package io.github.kreiseljustus.asmputils.core.data;
+
+public class ShopException extends Exception {
+    ShopException() {
+        super("Failed to create shop data");
+    }
+}
