+++ src/main/java/io/github/kreiseljustus/asmputils/core/data/ShopWaystoneUploadPacket.java
@@ -1,4 +1,4 @@
-package io.github.kreiseljustus.asmputils;
+package io.github.kreiseljustus.asmputils.core.data;
 
 import java.util.List;
