+++ src/main/java/io/github/kreiseljustus/asmputils/core/modules/shop/ServerValidator.java
@@ -1,7 +1,11 @@
-package io.github.kreiseljustus.asmputils;
+package io.github.kreiseljustus.asmputils.core.modules.shop;
 
 import com.google.common.reflect.TypeToken;
 import com.google.gson.Gson;
+import io.github.kreiseljustus.asmputils.Asmputils;
+import io.github.kreiseljustus.asmputils.ModConfig;
+import io.github.kreiseljustus.asmputils.core.Utils;
+import io.github.kreiseljustus.asmputils.core.data.ShopDataHolder;
 import org.jetbrains.annotations.NotNull;
 
 import java.io.BufferedReader;
